[REFACTOR/#170] item 패키지 분리 Enums
