+++ src/main/java/com/favoriteplace/app/item/service/ShopService.java
@@ -4,9 +4,9 @@ import static com.favoriteplace.global.exception.ErrorCode.ITEM_NOT_EXISTS;
 
 import com.favoriteplace.app.item.converter.ShopConverter;
 import com.favoriteplace.app.member.domain.Member;
-import com.favoriteplace.app.domain.enums.ItemCategory;
-import com.favoriteplace.app.domain.enums.ItemType;
-import com.favoriteplace.app.domain.enums.SaleStatus;
+import com.favoriteplace.app.item.domain.enums.ItemCategory;
+import com.favoriteplace.app.item.domain.enums.ItemType;
+import com.favoriteplace.app.item.domain.enums.SaleStatus;
 import com.favoriteplace.app.item.domain.AcquiredItem;
 import com.favoriteplace.app.item.domain.Item;
 import com.favoriteplace.app.item.controller.dto.ItemDto;
