+++ src/main/java/com/favoriteplace/app/service/MyPageQueryService.java
@@ -3,8 +3,8 @@ package com.favoriteplace.app.service;
 import com.favoriteplace.app.converter.MyPageConverter;
 import com.favoriteplace.app.domain.Block;
 import com.favoriteplace.app.member.domain.Member;
-import com.favoriteplace.app.domain.enums.ItemType;
-import com.favoriteplace.app.domain.enums.SaleStatus;
+import com.favoriteplace.app.item.domain.enums.ItemType;
+import com.favoriteplace.app.item.domain.enums.SaleStatus;
 import com.favoriteplace.app.item.domain.AcquiredItem;
 import com.favoriteplace.app.domain.travel.CompleteRally;
 import com.favoriteplace.app.domain.travel.LikedRally;
